+++ bindings/xcframework/merlin_transcript.xcframework/ios-arm64-simulator/merlin_transcript.framework/Headers/merlin-transcript.h
@@ -1,0 +1,4 @@
+// File automatically generated by swift-bridge.
+struct __private__ResultPtrAndPtr __swift_bridge__$ext_sign_raise_transcript(void* public_key, void* secret, void* message);
+
+
