+++ bindings/xcframework/merlin_transcript.xcframework/ios-arm64/merlin_transcript.framework/Modules/module.modulemap
@@ -1,0 +1,5 @@
+framework module merlin_transcript {
+    umbrella header "../Headers/merlin-transcript.h"
+    header "../Headers/SwiftBridgeCore.h"
+    export *
+}
