--- bindings/xcframework/merlin_transcript.xcframework/ios-arm64_x86_64-simulator/Headers/merlin-transcript.h
@@ -1,4 +1,0 @@
-// File automatically generated by swift-bridge.
-struct __private__ResultPtrAndPtr __swift_bridge__$ext_sign_raise_transcript(void* public_key, void* secret, void* message);
-
-
