--- bindings/xcframework/merlin_transcript.xcframework/ios-arm64_x86_64-simulator/Headers/module.modulemap
@@ -1,5 +1,0 @@
-module merlin_transcript {
-    header "SwiftBridgeCore.h"
-    header "merlin-transcript.h"
-    export *
-}
